Adds variables that are used in the jsons

+++ Monster.cpp
@@ -31,7 +31,7 @@
         return ATK;
     }
 
-    double const & Monster::getAttackCoolDown() const
+    float const & Monster::getAttackCoolDown() const
     {
         return ACD;
     }
